refactor: move utility fns up
